% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.functions.R
\name{integration}
\alias{integration}
\title{discrete integration of series or a function}
\usage{
integration(y, x = seq(f = from, t = to, len = n), from = min(x),
  to = max(x), n = length(x), method = "center")
}
\arguments{
\item{y}{The values to integrate. Either a numeric vector of function values 
corresponding to elements in \code{x} or a function that takes a single 
argument \code{x} and returns the corresponding value \code{y}.\cr
For example: \code{y = function(x) x ^ 2}.}

\item{x}{numeric vector of \code{y}-value-positions. Can be left unspecified
if \code{from}, \code{to} and \code{n} are given.}

\item{from}{lower bound of integration, will be used as smallest \code{x}. Unused if \code{x} is specified.}

\item{to}{upper bound of integration, will be used as highest \code{x}. Unused if \code{x} is specified.}

\item{n}{integer. Amount of steps / resolution of integration. Determines the
length of \code{x}. Unused if \code{x} is specified.}

\item{method}{integration method. Either \code{center},\code{left},\code{right},\code{upper} or \code{lower}.}
}
\value{
numeric value of integration
}
\description{
Numeric discrete integration of series or function \code{y} over variable \code{x}.
}

